split mira/catlab into different sub-services
