+++ packages/client/hmi-client/src/components/models/tera-stratified-value-matrix.vue
@@ -75,10 +75,7 @@
 <script setup lang="ts">
 import { ref, onMounted } from 'vue';
 import { cloneDeep, isEmpty } from 'lodash';
-import {
-	getCatlabStatesMatrixData,
-	getCatlabTransitionsMatrixData
-} from '@/model-representation/petrinet/petrinet-service';
+import { getAMRPresentationData } from '@/model-representation/petrinet/catlab-petri';
 import { createMatrix1D, createMatrix2D } from '@/utils/pivot';
 import Dropdown from 'primevue/dropdown';
 import { Initial, ModelConfiguration, ModelParameter, Rate } from '@/types/Types';
@@ -159,15 +156,13 @@ function updateModelConfigValue(variableName: string) {
 function configureMatrix() {
 	console.log(props.modelConfiguration.configuration);
 
+	const result = getAMRPresentationData(props.modelConfiguration.configuration);
+
 	// Get only the states/transitions that are mapped to the base model
 	const matrixData =
 		props.nodeType === NodeType.State
-			? getCatlabStatesMatrixData(props.modelConfiguration.configuration).filter(
-					(d) => d['@base'] === props.id
-			  )
-			: getCatlabTransitionsMatrixData(props.modelConfiguration.configuration).filter(
-					(d) => d['@base'] === props.id
-			  );
+			? result.stateMatrixData.filter((d) => d._base === props.id)
+			: result.transitionMatrixData.filter((d) => d._base === props.id);
 
 	console.log(matrixData);
 
@@ -175,8 +170,8 @@ function configureMatrix() {
 
 	// Grab dimension names from the first matrix row
 	const dimensions = [cloneDeep(matrixData)[0]].map((d) => {
-		delete d.id;
-		delete d['@base'];
+		delete d._id;
+		delete d._base;
 		return Object.keys(d);
 	})[0];
